refactor: shared => common
