+++ website/prorata/pages/index.tsx
@@ -3,7 +3,7 @@ import { NextPage } from 'next'
 import Head from 'next/head'
 import { Flex, useColorModeValue } from '@chakra-ui/react'
 
-import { AllocationRequest, AllocationResponse } from '../shared'
+import { AllocationRequest, AllocationResponse } from '../common'
 import { fetch, useAsyncState } from '../frontend'
 import { Prorata } from '../components'
